backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
GeoAlchemy2==0.14.2
Shapely==2.0.1
redis==5.0.0
celery==5.3.4
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
bcrypt==4.0.1
Pillow==10.0.0
requests==2.31.0
firebase-admin==6.2.0
stripe==7.0.0
pandas==2.1.1
numpy==1.24.3
geopy==2.4.0
folium==0.14.0
gunicorn==21.2.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
